Add files via upload
The Online shopping got its boost at the time of covid. The basic mechanism of the project flows with: registration  Login  main_page  View Products Add them to Cart  Buy Them. The technology used for the project is enterprize java ( Using Servlets, Jsp , JDBC and bootstrap or graphic user interface). This project includes validation of user and data management  In this project I have performed updation and insertion of data from database , Session and redirect methods are  also implemented in this project which make it easy to use and manage .